+++ ChangeLog.md
@@ -29,7 +29,7 @@
 
 **2023-09-18**
 - 🌟新增对话修改功能【右键菜单】
-- 🌟新增工具使用教程<a href='https://ele-cat.gitee.io/tools/Vue3WechatTool.html#使用教程' target="_blank">使用教程</a>
+- 🌟新增工具使用教程<a href='https://ele-cat.github.io/tools/Vue3WechatTool.html#使用教程' target="_blank">使用教程</a>
 - 🐞修复一系列系统遗留问题
 
 **2023-09-17**